requirements.txt
nats-py~=2.10.0
